Extract more of the keyboard finding logic out of the react component
* not that we're putting it somewhere good, but at least the component
  is cleaner
